kucoin: convert to TS, disable L3updates
